backend_sig/requirements.txt
Flask
flask-cors
flask-pymongo
pymongo
PyJWT
python-dotenv
Werkzeug
